CAMEL-21625: camel-jbang - camel k8s run may not cleanup properly on openshift
